requirements.txt
fastapi
uvicorn
motor
loguru
httpx
playwright
anthropic
PyJWT
apscheduler
